fix: extract group command guard to satisfy backend-size limit
Move unknown-subcommand guard helpers out of root.go so pkg/cli/root.go stays within the 1000-line backend-size cap after the factory save removal.
